created separate module for rebounds in simulator
